+++ src/dd-resizable.ts
@@ -237,14 +237,11 @@ export class DDResizable extends DDBaseImplement implements HTMLElementExtendOpt
   /** @internal */
   protected _getChange(event: MouseEvent, dir: string): Rect {
     const oEvent = this.startEvent;
-    const containerElement = Utils.getPositionContainerElement(this.el.parentElement);
-    const containerRect = containerElement.getBoundingClientRect();
-
     const newRect = { // Note: originalRect is a complex object, not a simple Rect, so copy out.
       width: this.originalRect.width,
       height: this.originalRect.height + this.scrolled,
-      left: this.originalRect.left - containerRect.left,
-      top: this.originalRect.top - this.scrolled - containerRect.top
+      left: this.originalRect.left,
+      top: this.originalRect.top - this.scrolled
     };
 
     const offsetX = event.clientX - oEvent.clientX;
@@ -280,12 +277,10 @@ export class DDResizable extends DDBaseImplement implements HTMLElementExtendOpt
 
   /** @internal constrain the size to the set min/max values */
   protected _constrainSize(oWidth: number, oHeight: number): Size {
-    const { scaleX, scaleY } = Utils.getScaleForElement(this.el);
-    const o = this.option;
-    const maxWidth = o.maxWidth ? o.maxWidth * scaleX : Number.MAX_SAFE_INTEGER;
-    const minWidth = o.minWidth ? o.minWidth * scaleX : oWidth;
-    const maxHeight = o.maxHeight ? o.maxHeight * scaleY : Number.MAX_SAFE_INTEGER;
-    const minHeight = o.minHeight ? o.minHeight * scaleY : oHeight;
+    const maxWidth = this.option.maxWidth || Number.MAX_SAFE_INTEGER;
+    const minWidth = this.option.minWidth || oWidth;
+    const maxHeight = this.option.maxHeight || Number.MAX_SAFE_INTEGER;
+    const minHeight = this.option.minHeight || oHeight;
     const width = Math.min(maxWidth, Math.max(minWidth, oWidth));
     const height = Math.min(maxHeight, Math.max(minHeight, oHeight));
     return { width, height };
@@ -293,12 +288,17 @@ export class DDResizable extends DDBaseImplement implements HTMLElementExtendOpt
 
   /** @internal */
   protected _applyChange(): DDResizable {
+    let containmentRect = { left: 0, top: 0, width: 0, height: 0 };
+    if (this.el.style.position === 'absolute') {
+      const containmentEl = this.el.parentElement;
+      const { left, top } = containmentEl.getBoundingClientRect();
+      containmentRect = { left, top, width: 0, height: 0 };
+    }
     if (!this.temporalRect) return this;
-    const { scaleX, scaleY } = Utils.getScaleForElement(this.el);
-    this.el.style.width = `${Math.round(this.temporalRect.width / scaleX)}px`;
-    this.el.style.height = `${Math.round(this.temporalRect.height / scaleY)}px`;
-    this.el.style.top = `${Math.round(this.temporalRect.top / scaleY)}px`;
-    this.el.style.left = `${Math.round(this.temporalRect.left / scaleX)}px`;
+    Object.keys(this.temporalRect).forEach(key => {
+      const value = this.temporalRect[key];
+      this.el.style[key] = value - containmentRect[key] + 'px';
+    });
     return this;
   }
 
@@ -311,22 +311,23 @@ export class DDResizable extends DDBaseImplement implements HTMLElementExtendOpt
 
   /** @internal */
   protected _ui = (): DDUIData => {
-    const { scaleX, scaleY } = Utils.getScaleForElement(this.el);
+    const containmentEl = this.el.parentElement;
+    const containmentRect = containmentEl.getBoundingClientRect();
     const newRect = { // Note: originalRect is a complex object, not a simple Rect, so copy out.
       width: this.originalRect.width,
-      height: (this.originalRect.height + this.scrolled),
+      height: this.originalRect.height + this.scrolled,
       left: this.originalRect.left,
-      top: (this.originalRect.top - this.scrolled)
+      top: this.originalRect.top - this.scrolled
     };
     const rect = this.temporalRect || newRect;
     return {
       position: {
-        left: rect.left / scaleX,
-        top: rect.top / scaleY,
+        left: rect.left - containmentRect.left,
+        top: rect.top - containmentRect.top
       },
       size: {
-        width: rect.width / scaleX,
-        height: rect.height / scaleY,
+        width: rect.width,
+        height: rect.height
       }
       /* Gridstack ONLY needs position set above... keep around in case.
       element: [this.el], // The object representing the element to be resized